+++ CHANGELOG.md
@@ -13,6 +13,7 @@ For a full diff see [`0.6.1...main`][0.6.1...main].
 
 - Added form to enter reading progress directly in the book details modal
 - Added option to sort books by reading progress
+- Added pages to book metadata and views
 
 ### Changed